refactor(handler): update package configuration
