+++ README.md
@@ -11,3 +11,6 @@ GOOS=linux GOARCH=amd64 go build -o main
 zip handler.zip ./main
 mv handler.zip ~/Desktop
 ```
+
+## Fix format command
+`go fmt ./..`
