+++ src/handler/listWork.go
@@ -1,0 +1,39 @@
+package handler
+
+import (
+	"errors"
+	"fmt"
+
+	"github.com/is09-souzou/AppSync-Resolver-Mapping-Lambda/src/model"
+	"github.com/is09-souzou/AppSync-Resolver-Mapping-Lambda/src/types"
+)
+
+// ListWorkHandle List Work Handle
+func ListWorkHandle(arg WorkDelete, identity types.Identity) (Work, error) {
+
+	svc, err := model.GetSVC()
+
+	if err != nil {
+		return Work{}, err
+	}
+
+	work, err := model.GetWorkByID(svc, arg.ID)
+
+	if err != nil {
+		return Work{}, err
+	}
+
+	if work.UserID != identity.Sub {
+		return Work{}, errors.New("Can delete only oneself")
+	}
+
+	err = model.DeleteWorkByID(svc, arg.ID)
+
+	if err != nil {
+		fmt.Println("Got error calling DeleteUserHandle:")
+		fmt.Println(err.Error())
+		return Work{}, err
+	}
+
+	return Work{ID: arg.ID}, nil
+}
